requirements.txt
google-cloud-bigquery
google-cloud-storage
google-cloud-secret-manager
pyyaml
requests
google-auth